Added the ability to load modules on tiles from XML, show modules on tiles to the user, and save modules on tiles to XML. Recursive modules are not yet supported; nor are tiles with multiple modules on them.

+++ sp_java_next/advances/ExampleUnit.java
@@ -164,7 +164,7 @@ public class ExampleUnit implements Module, Serializable {
 	 * @return the unit's UUID
 	 */
 	@Override
-	public int getUuid() {
+	public long getUuid() {
 		throw new IllegalStateException("Unimplemented");
 	}
 }
